--- app/components/navigator/navigator.js
@@ -1,1 +1,0 @@
-console.log("navigatior");
