--- app/components/screens/Reservation.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-import {View, Text} from "react-native";
-
-function Reservation() {
-	return (
-		<View>
-			<Text>예약 화면입니다.</Text>
-		</View>
-	);
-}
-
-export default Reservation;
