+++ app/components/screens/Reservation/ReservingProcess.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import {View, Text} from 'react-native';
+
+function ReservingProcess() {
+    return (
+        <View>
+            <Text>{"실제 예약을 진행하는 화면입니다."}</Text>
+        </View>
+    );
+}
+
+export default ReservingProcess;
